refactor: [WLEO-425] Refactored matchRequestToClaims method to make it more readable and created moved it to `features/wallet/utils/proximity.ts` replacing the old `useProximity` hook utility methods
