+++ ts/features/wallet/utils/proximity.ts
@@ -1,114 +1,170 @@
-import {
-  VerifierRequest,
-  AcceptedFields
-} from '@pagopa/io-react-native-proximity';
-import {wellKnownCredential} from './credentials';
-import {StoredCredential} from './types';
+import {CBOR} from '@pagopa/io-react-native-cbor';
+import {VerifierRequest} from '@pagopa/io-react-native-proximity';
+import {ParsedCredential, StoredCredential} from './types';
 
-/**
- * This function generates the accepted fields for the VerifierRequest and sets each requested field to true.
- * WARNING: This function is a quick and dirty implementation copied from the proximity example app. It will be replaced in the future.
- * @param request - The request object containing the requested fields
- * @returns A new object with the same structure as the request, but with all values set to true
- */
-export const generateAcceptedFields = (
-  request: VerifierRequest['request']
-): AcceptedFields => {
-  // Cycle through the requested credentials
-  const result: AcceptedFields = {};
-  // eslint-disable-next-line guard-for-in
-  for (const credentialKey in request) {
-    const credential = request[credentialKey];
-    if (!credential) {
-      continue;
-    }
-
-    // Cycle through the requested namespaces and the isAuthenticated field
-    const namespaces: AcceptedFields['credential'] = {};
-    for (const namespaceKey in credential) {
-      // Skip the isAuthenticated field
-      if (!credential[namespaceKey] || namespaceKey === 'isAuthenticated') {
-        continue;
-      }
-
-      // Cycle through the requested fields and set them to true
-      const fields: AcceptedFields['credential']['namespace'] = {};
-      // eslint-disable-next-line guard-for-in
-      for (const fieldKey in credential[namespaceKey]!) {
-        // eslint-disable-next-line functional/immutable-data
-        fields[fieldKey] = true;
-      }
-      // eslint-disable-next-line functional/immutable-data
-      namespaces[namespaceKey] = fields;
-    }
-    // eslint-disable-next-line functional/immutable-data
-    result[credentialKey] = namespaces;
-  }
-
-  return result;
-};
-
-export enum RequestType {
-  MDL,
-  HEALTHID,
-  BOTH
-}
-
-/**
- * Utility funciton to check the request type based on the keys of the request object.
- * It currently supports only mDL and HealthID credentials or both.
- * @param requestKeys - The keys of the request object which contains the credential type
- * @returns A RequestType enum value indicating the type of request
- * @throws Error if the request contains multiple keys or if the key is not the mDL credential type
- */
-export const getTypeRequest = (requestKeys: Array<string>) => {
-  if (
-    requestKeys.length === 1 ||
-    (requestKeys.length === 2 &&
-      requestKeys.every(
-        key =>
-          key === wellKnownCredential.DRIVING_LICENSE ||
-          key === wellKnownCredential.HEALTHID
-      ))
-  ) {
-    if (requestKeys.length === 2) {
-      return RequestType.BOTH;
-    } else if (requestKeys[0] === wellKnownCredential.HEALTHID) {
-      return RequestType.HEALTHID;
-    } else if (requestKeys[0] === wellKnownCredential.DRIVING_LICENSE) {
-      return RequestType.MDL;
-    }
-  }
-
-  throw new Error('Unexpected request keys. Expected only mDL or HealthID.');
+type StoredCredentialWithIssuerSigned = StoredCredential & {
+  issuerSigned: CBOR.IssuerSigned;
 };
 
 /**
- * Utility functions which filters the documents based on the request type. Currently it supports only mDL and HealthID credentials or both.
- * @param type - The type of request returned by the `getTypeRequest` function
- * @param documents - The array of stored credentials to filter
- * @returns An array of stored credentials that match the request type
+ * Helper function to generate the attributes for the {@link matchRequestToClaims} method.
+ * For a given Verifier Request namespace-required attributes, this method looks for matches
+ * and creates an object where to each namespace corresponds an object containing the
+ * attribute's value and display info as found in the {@link StoredCredentialWithIssuerSigned}
  */
-export const getDocumentsByRequestType = (
-  type: RequestType,
-  documents: Array<StoredCredential>
+const attributesReducerGenerator =
+  (credential: StoredCredentialWithIssuerSigned) =>
+  (prev: Record<string, ParsedCredential[string]>, attribute: string) => {
+    // If the credential contains the attribute, add it to the accumulator
+    if (credential.parsedCredential[attribute]) {
+      return {
+        ...prev,
+        [attribute]: credential.parsedCredential[attribute]
+      };
+    }
+    // Go to the next attribute otherwise
+    return {...prev};
+  };
+
+/**
+ * Helper function to generate the namespaces for the {@link matchRequestToClaims} method.
+ * For a given credential and Verifier required namespaces, this method looks for matches
+ * and creates an object where to each namespace corresponds an object mapping the attributes to 
+ * their value and display info as found in the {@link StoredCredentialWithIssuerSigned}
+
+ * containing the attribute value and display info as found in the {@link StoredCredentialWithIssuerSigned}
+ 
+ * @param credential a {@link StoredCredentialWithIssuerSigned} containing the claims values and display info
+ * @returns an object where to each namespace corresponds an object mapping the attributes to 
+ *          their value and display info as found in the {@link StoredCredentialWithIssuerSigned}
+ */
+const nameSpaceReducerGenerator =
+  (credential: StoredCredentialWithIssuerSigned) =>
+  /**
+   * @param namespace : A Verifier Request namespace
+   * @param attributes : The Verifier requested namespace attributes
+   */
+  (
+    accumulator: Record<string, Record<string, ParsedCredential[string]>>,
+    [namespace, attributes]: [string, boolean | Record<string, boolean>]
+  ) => {
+    // Check first if the credential contains the required namespace
+    const foundNamespace = Object.entries(
+      credential.issuerSigned.nameSpaces
+    ).find(([ns]) => ns === namespace);
+    if (foundNamespace) {
+      // If found, combine the attributes required for the specific namespace
+      // by the Verifier Request with claim values and display info from the credentials,
+      // if matches are found
+      const foundAttributes = Object.keys(attributes).reduce(
+        attributesReducerGenerator(credential),
+        {} as Record<string, ParsedCredential[string]>
+      );
+      if (Object.keys(foundAttributes).length !== 0) {
+        // If attribute have been found add the namespace new entry containing attribute
+        // values and display data to the accumulator
+        return {
+          ...accumulator,
+          [namespace]: foundAttributes
+        };
+      } else {
+        // If no attribute has been found go to the next namespace
+        return {
+          ...accumulator
+        };
+      }
+    }
+    // If the credential doesn't contain the namespace go to the next one
+    return {...accumulator};
+  };
+
+/**
+ * This method matches every required attribute contained in a {@link VerifierRequest} to its
+ * corresponding entry in one of the decoded credentials and generates an object which, for
+ * every matched attribute, corresponds the claim value and display info found inside the credential,
+ * keeping the original
+ * {
+ *  credentialType : {
+ *    namespace : {
+ *      attributes...
+ *    }
+ *  }
+ * }
+ * structure.
+ * @param request A {@link VerifierRequest} request field
+ * @param decodedCredentials An array of {@link StoredCredentialWithIssuerSigned} mDoc credentials
+ * @returns see description
+ */
+const mapVerifierRequestToClaimInfo = (
+  request: VerifierRequest['request'],
+  decodedCredentials: Array<StoredCredentialWithIssuerSigned>
+) =>
+  Object.entries(request).reduce((accumulator, [credentialType, value]) => {
+    // Search for a credential of the same credential type specified in the Verifier Request
+    const credential = decodedCredentials.find(
+      cred => cred.credentialType === credentialType
+    );
+    if (credential) {
+      // Key : namespace
+      // Value : attributes
+      const foundNamespaces = Object.entries(value)
+        // Filter the isAuthenticated key, which is not a namespace
+        .filter(([key, _]) => key !== 'isAuthenticated')
+        // Combine the namespaces and attributes required by the Verifier Request with
+        // claim values and display info from the credentials, if matches are found
+        .reduce(
+          nameSpaceReducerGenerator(credential),
+          {} as Record<string, Record<string, ParsedCredential[string]>>
+        );
+      if (Object.keys(foundNamespaces).length === 0) {
+        // If no namespace has been found or no attributes in any namespace have been found go to the next Verifier required credential
+        return {...accumulator};
+      } else {
+        // Otherwise add a new entry to the accumulator
+        return {
+          ...accumulator,
+          [credentialType]: foundNamespaces
+        };
+      }
+    }
+    // If no matching credential has been found go to the next Verifier required credential
+    return {
+      ...accumulator
+    };
+  }, {} as Record<string, Record<string, Record<string, ParsedCredential[string]>>>);
+
+/**
+ * This helper function takes a {@link VerifierRequest}, looks for
+ * the presence of the required claims in the mDoc credentials and, if
+ * found adds the {@link ParsedCredential} entry for the attribute to
+ * an object following the same path structure of the original credential
+ * @param verifierRequest The authentication request sent by the verifier
+ * @param credentialsMdoc The mdoc credentials contained in the wallet
+ * @returns An object that is a record of credential types to an object
+ * which has the same structure of a decoded mDoc credential's namespaces
+ * except for the fact that the namespace attribute keys are mapped to
+ * the value corresponding to the same attribute key inside of the
+ * {@link ParsedCredential} object.
+ */
+export const matchRequestToClaims = async (
+  verifierRequest: VerifierRequest,
+  credentialsMdoc: Array<StoredCredential>
 ) => {
-  switch (type) {
-    case RequestType.MDL:
-      return documents.filter(
-        doc => doc.credentialType === wellKnownCredential.DRIVING_LICENSE
-      );
-    case RequestType.HEALTHID:
-      return documents.filter(
-        doc => doc.credentialType === wellKnownCredential.HEALTHID
-      );
-    case RequestType.BOTH:
-      return documents.filter(
-        doc =>
-          doc.credentialType === wellKnownCredential.DRIVING_LICENSE ||
-          doc.credentialType === wellKnownCredential.HEALTHID
-      );
-    default:
-      throw new Error('Unexpected request type');
-  }
+  const decodedCredentials: Array<StoredCredentialWithIssuerSigned> =
+    await Promise.all(
+      credentialsMdoc.map(async credential => {
+        const decodedIssuerSigned = await CBOR.decodeIssuerSigned(
+          credential.credential
+        );
+        return {
+          ...credential,
+          issuerSigned: decodedIssuerSigned
+        };
+      })
+    );
+
+  return mapVerifierRequestToClaimInfo(
+    verifierRequest.request,
+    decodedCredentials
+  );
 };
